# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/learning/test": {
      "get": {
        "operationId": "LearningController_test",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/levels": {
      "get": {
        "operationId": "LearningController_getLevels",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LevelItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/lessons": {
      "get": {
        "operationId": "LearningController_getLessons",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "levelId",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LessonItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPageResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/lessons/{id}": {
      "get": {
        "operationId": "LearningController_getLessonDetail",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/LessonDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/topics": {
      "get": {
        "operationId": "LearningController_getTopics",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "lessonId",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TopicItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPageResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/learning/stories": {
      "get": {
        "operationId": "LearningController_getStories",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "levelId",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/StoryItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPageResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/levels": {
      "get": {
        "operationId": "LevelsController_getLevels",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LevelItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Levels"
        ]
      }
    },
    "/api/v1/lessons": {
      "get": {
        "operationId": "LessonsController_getLessons",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "levelId",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LessonItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPageResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/v1/lessons/{id}": {
      "get": {
        "operationId": "LessonsController_getLessonDetail",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/LessonDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/v1/topics": {
      "get": {
        "operationId": "TopicsController_getTopics",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "lessonId",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TopicItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPageResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Topics"
        ]
      }
    },
    "/api/v1/stories": {
      "get": {
        "operationId": "StoriesController_getStories",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "levelId",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/StoryItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPageResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Stories"
        ]
      }
    },
...
  },
  "info": {
    "title": "HSK System API",
    "description": "Generated from the backend code. Envelope, errors and request id: docs/api/api.md §1.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LevelItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "orderIndex": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "orderIndex"
        ]
      },
      "LessonItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "orderIndex": {
            "type": "number"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "orderIndex",
          "slug"
        ]
      },
      "LessonDetailLevelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "orderIndex": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "orderIndex"
        ]
      },
      "LessonDetailTopicDto": {
        "type": "object",
        "properties": {
          "content": {
            "oneOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {}
              }
            ]
          },
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "orderIndex": {
            "type": "number"
          }
        },
        "required": [
          "content",
          "id",
          "title",
          "orderIndex"
        ]
      },
      "LessonDetailWordMeaningDto": {
        "type": "object",
        "properties": {
          "en": {
            "type": "string",
            "nullable": true
          },
          "vi": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "en",
          "vi"
        ]
      },
      "LessonDetailWordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "hanzi": {
            "type": "string"
          },
          "traditional": {
            "type": "string",
            "nullable": true
          },
          "pinyin": {
            "type": "string"
          },
          "pinyinTone": {
            "type": "string",
            "nullable": true
          },
          "meanings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonDetailWordMeaningDto"
            }
          }
        },
        "required": [
          "id",
          "hanzi",
          "traditional",
          "pinyin",
          "pinyinTone",
          "meanings"
        ]
      },
      "LessonDetailStoryDto": {
        "type": "object",
        "properties": {
          "content": {
            "oneOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {}
              }
            ]
          },
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "content",
          "id",
          "title",
          "slug"
        ]
      },
      "LessonDetailExerciseMediaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "url": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "url",
          "type",
          "mimeType",
          "duration"
        ]
      },
      "LessonDetailExerciseDto": {
        "type": "object",
        "properties": {
          "content": {
            "oneOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {}
              }
            ]
          },
          "id": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "orderIndex": {
            "type": "number"
          },
          "media": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LessonDetailExerciseMediaDto"
              }
            ]
          }
        },
        "required": [
          "content",
          "id",
          "type",
          "prompt",
          "version",
          "orderIndex",
          "media"
        ]
      },
      "LessonDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/LessonDetailLevelDto"
          },
          "topics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonDetailTopicDto"
            }
          },
          "words": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonDetailWordDto"
            }
          },
          "stories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonDetailStoryDto"
            }
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonDetailExerciseDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "level",
          "topics",
          "words",
          "stories",
          "exercises"
        ]
      },
      "TopicItemDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "id": {
            "type": "number"
          },
          "lessonId": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "orderIndex": {
            "type": "number"
          }
        },
        "required": [
          "content",
          "id",
          "lessonId",
          "title",
          "orderIndex"
        ]
      },
      "StoryItemDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "id": {
            "type": "number"
          },
          "levelId": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "content",
          "id",
          "levelId",
          "title",
          "slug"
        ]
      },
...
      "ApiResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "description": "Same value as the `X-Request-ID` response header."
          }
        },
        "required": [
          "timestamp",
          "requestId"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "ApiPageResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "description": "Same value as the `X-Request-ID` response header."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "timestamp",
          "requestId",
          "pagination"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, for example `UNAUTHORIZED`."
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResponseMetaDto"
          }
        },
        "required": [
          "success",
          "error",
          "meta"
        ]
      }
    },
    "parameters": {
      "RequestId": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "UUID or 32 hex characters; any other value is replaced by a new UUID. Echoed in the X-Request-ID response header and meta.requestId.",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "ErrorEnvelope": {
        "description": "Error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
            }
          }
        }
      }
    }
  }
}